better resolution
SCCS-file: date.c
SCCS-SID: 7.14

+++ date.c
@@ -96,7 +96,7 @@ char *		argv[];
 	register int		tflag = 0;
 	register int		minuteswest;
 	register int		dsttime;
-	register float		adjust;
+	register double		adjust;
 	time_t			now;
 	time_t			t;
